requirements.txt
opencv-python
numpy
PyQt6
mediapipe
pymongo
pyttsx3
SpeechRecognition
pyaudio
google-genai
